Remove the first landing page.
